Add tests on djangoes.ConnectionHandler
Testing the default behavior is the foundation to build a better ConnectionHandler and have better testing.
It helps to discover what are the best default settings, and to define the minimal set of information required.

It is not as good as a documentation when it comes to learning how to use djangoes, but it will be a good starting point.