Add a djangoes test runner for Django projects
In order to write and run tests in a django projects using ElasticSearch, djangoes now provide a test runner, subclassing the Django test runner.

When the test suite starts, each server and index configuration is verified and the live settings are replaced by the test settings.

And this time thanks to TDD, the runner and the setup function are fully tested.