Add py.test plugin for djangoes
A popular way to run tests in django is to use py.test: djangoes now provides a plugin, using the setup entry point for that purpose, and reuse function from the py.test plugin pytest-django.

Of course, djangoes still works without these dependencies, but they are required when testing with py.test.