Fix params used as positional argument.
With the elasticsearch library, the "params" positional argument is not meant to be used "as-is", as it will be created by the decorator `query_params`.
So when calling a method of the djangoes backend, an error was raised saying `params` is given twice. This fixes the issue by using the **keywords arguments.

Also, hotfix with bump version to 0.1.4.