Use module.backend.Class instead of module.backend
After discussions and reading about backend configuration, it's better to use a configuration with `backend.Class` than `backend` expecting an attribute `Class` of this module. In fact, it looks like the database layer of Django is one of the latest part of Django to use the module path, as a legacy from its past.

Anyway, tests have been updated, and the version bumped to an alpha release for 0.2.0 (alpha 1).