Refactor and improve djangoes.backends module.
The `djangoes.backends` module became messy and needed improvements - and also, more documentation.
In this commit, the initial base classes are now in `djangoes.backends.abstracts`, and the `djangoes.backends.elasticsearch` now contains only the backend implementations that rely on the official `elastcisearch-py` library.

The documentation has been updated with more information about backends, with links to the official documentation. It's not perfect nor complete yet, as many information are still missing (for example, the exact list of arguments to use in the `PARAMS` key of connection's configuration dict).

Tests have been updated accordingly.

And because of all these changes, I run into the next version 0.3.0 - first dev version.