Remove the meta attribute from backends
It appears that `connection.meta` is not good enough, and is over
engineered at the moment: features can be accessed through
`connection.client` for now.

I prefer to focus on testability, and to provide better documentation,
than support a "meta" interface I'm not sure to keep up to date.

Simplicity first.