Add SETTINGS to create indices in tests
In tests, we need to create indices *with* their full configuration,
otherwise, Djangoes's test cases would be useless.

The idea is to have a `SETTINGS` key in the index configuration dict
and then to use the new method of backends `get_indices_with_settings`
in test case methods.

The documentation has been updated, and tests too. Version is now
0.3.0.dev2.