Replace mixin test static class methods by methods
We can not ensure that all runners (in particular py.test) will setup
djangoes environment at the `setUpClass` stage, so allowing such
methods to be called as static/class methods could be more dangerous
than useful.

Something has to be done for that, but it requires more thought to get
a smart and simple solution for everyone.