Initial IndexHandler implementation
This is the starting point: I want to do that:

```
>>> from djangoes import indices
>>> es_index = indices['default']
>>> es_index.search(body)
{... standard ES search result ...}
```

So, the first step is to have a basic `IndexHandler`, which will handle
the configuration from `settings.ES_INDICES`.

There is a default index now, using the `default` alias. That's all at
the moment on the `IndexHandler` side.

On the `ConnectionHandler` side, it will slowly be transformed into a
connection only handler - no indice management.

Also, without modifying any tests of `djangoes.connections`, it now
comes from the default `IndexHandler`. This is a good start for a
soft transition.